+++ frontend/src/services/userServices.js
@@ -1,43 +1,30 @@
 
 import API from "./api";
-import { handleError } from "../utils/handleError";
 
 // Get all users
 export const getUsers = async () => {
-  try {
-    const res = await API.get("/users");
-    return res.data;
-  } catch (err) {
-    throw handleError(err);
-  }
+      const res = await API.get("/users");
+      return res.data;
+  
 };
 
 // Get user by ID
 export const getUserById = async (id) => {
-  try {
+  if(!id) throw new Error("User ID is required");
     const res = await API.get(`/users/${id}`);
     return res.data;
-  } catch (err) {
-    throw handleError(err);
-  }
 };
 
 // Update user
 export const updateUser = async (id, userData) => {
-  try {
-    const res = await API.put(`/users/${id}`, userData);
+    if(!id) throw new Error("User ID is required");
+      const res = await API.put(`/users/${id}`, userData);
     return res.data;
-  } catch (err) {
-    throw handleError(err);
-  }
 };
 
 // Delete user
 export const deleteUser = async (id) => {
-  try {
+   if(!id) throw new Error("User ID is required");
     const res = await API.delete(`/users/${id}`);
     return res.data;
-  } catch (err) {
-    throw handleError(err);
-  }
 };
